Add linear search algorithm
Implemented the linearSearchAlgorithm function that performs a linear search on an array to find a specific element. The function iterates through the array and compares each element with the requested value. If a match is found, the index of the first occurrence is returned; otherwise, ELEMENT_NOT_FOUND is returned.

Function signature:
uint32 linearSearchAlgorithm(uint32 data[], uint32 size, uint32 reqData)

Closes #123

Signed-off-by: MasameEh <samehms116@gmail.com>